Add Exeptions to services and handel them in the controllers.

+++ src/Footballize/Footballize.Services/Data/RecruitmentService.cs
@@ -1,8 +1,10 @@
 ﻿namespace Footballize.Services.Data
 {
     using System.Collections.Generic;
+    using System.Drawing;
     using System.Linq;
     using System.Threading.Tasks;
+    using Exceptions;
     using Footballize.Data.Repositories;
     using Mapping;
     using Microsoft.EntityFrameworkCore;
@@ -40,50 +42,75 @@
                 .SingleOrDefault();
         }
 
+        public Recruitment GetRecruitmentWithPlayers(string id)
+        {
+            return this.recruitmentRepository
+                .All()
+                .Include(x => x.RecruitedUsers)
+                .SingleOrDefault(x => x.Id == id);
+        }
+
         public async Task AddRecruitmentAsync(Recruitment recruitment)
         {
+            if (recruitment == null)
+            {
+                throw new ServiceException(ServiceException.InvalidRequestParameters);
+            }
+
             await this.recruitmentRepository.AddAsync(recruitment);
             await this.recruitmentRepository.SaveChangesAsync();
         }
 
-        public async Task LeaveRecruitmentAsync(string recruitmentId, string userId)
+        public async Task LeaveRecruitmentAsync(Recruitment recruitment, string userId)
         {
-            var game = this.recruitmentRepository
-                .All()
-                .Include(x => x.RecruitedUsers)
-                .SingleOrDefault(x => x.Id == recruitmentId);
+            if (recruitment == null)
+            {
+                throw new ServiceException(ServiceException.InvalidRequestParameters);
+            }
 
-            var gameUser = game?.RecruitedUsers.SingleOrDefault(user => user.UserId.Equals(userId));
+            if (recruitment.Status != GameStatus.Registration)
+            {
+                throw new ServiceException(ServiceException.KickPlayerOnlyInRegistrationMode);
+            }
+
+            var gameUser = recruitment?.RecruitedUsers.SingleOrDefault(u => u.UserId.Equals(userId));
 
             if (gameUser == null)
             {
-                return;
+                throw new ServiceException(ServiceException.InvalidRequestParameters);
             }
 
             this.recruiterUserRepository.Delete(gameUser);
 
-            game.RecruitedUsers.Remove(gameUser);
-            this.recruitmentRepository.Update(game);
+            recruitment.RecruitedUsers.Remove(gameUser);
+            this.recruitmentRepository.Update(recruitment);
             await this.recruitmentRepository.SaveChangesAsync();
+            await this.recruiterUserRepository.SaveChangesAsync();
         }
 
-        public async Task EnrollRecruitmentAsync(string recruitmentId, string userId)
+        public async Task EnrollRecruitmentAsync(Recruitment recruitment, User user)
         {
-            var gameToEnroll = await this.recruitmentRepository.GetByIdAsync(recruitmentId);
-            var enrolledUser = await this.userRepository.GetByIdAsync(userId);
 
-            if (gameToEnroll == null || enrolledUser == null)
+            if (user == null || recruitment == null)
+                throw new ServiceException(ServiceException.InvalidRequestParameters);
+
+            if (recruitment.Status != GameStatus.Registration || recruitment.RecruitedUsers.Count >= recruitment.MaximumPlayers)
             {
-                return;
+                throw new ServiceException(ServiceException.NotInRegistrationOrNoFreeSlot);
+            }
+
+            if (recruitment.RecruitedUsers.Any(x => x.UserId == user.Id))
+            {
+                throw new ServiceException(ServiceException.AlreadyJoined);
             }
 
             var enrolledGame = new RecruitmentUser
             {
-                User = enrolledUser,
-                Recruitment = gameToEnroll
+                User = user,
+                Recruitment = recruitment
             };
 
-            gameToEnroll.RecruitedUsers.Add(enrolledGame);
+            recruitment.RecruitedUsers.Add(enrolledGame);
 
             await this.recruiterUserRepository.AddAsync(enrolledGame);
             await this.recruiterUserRepository.SaveChangesAsync();
@@ -93,6 +120,22 @@
         public async Task StartRecruitmentAsync(string id)
         {
             var gameToStart = await this.recruitmentRepository.GetByIdAsync(id);
+
+            if (gameToStart == null)
+            {
+                throw new ServiceException(ServiceException.InvalidRequestParameters);
+            }
+
+            if (gameToStart.RecruitedUsers.Count != gameToStart.MaximumPlayers)
+            {
+                throw new ServiceException(ServiceException.RequiredNumberOfPlayersNotReached);
+            }
+
+            if (gameToStart.Status != GameStatus.Registration)
+            {
+                throw new ServiceException(ServiceException.ThisGameIsAlreadyStarted);
+            }
+
             gameToStart.Status = GameStatus.Started;
 
             this.recruitmentRepository.Update(gameToStart);
@@ -101,18 +144,40 @@
 
         public async Task CompleteRecruitmentAsync(string id)
         {
-            var gameToStart = await this.recruitmentRepository.GetByIdAsync(id);
-            gameToStart.Status = GameStatus.Finished;
+            var game = await this.recruitmentRepository.GetByIdAsync(id);
 
-            this.recruitmentRepository.Update(gameToStart);
+            if (game == null)
+            {
+                throw new ServiceException(ServiceException.InvalidRequestParameters);
+            }
+
+            if (game.Status != GameStatus.Started)
+            {
+                throw new ServiceException(ServiceException.ThisGameIsNotStartedYet);
+            }
+
+            game.Status = GameStatus.Finished;
+
+            this.recruitmentRepository.Update(game);
             await this.recruitmentRepository.SaveChangesAsync();
         }
 
         public async Task DeleteRecruitmentAsync(string id)
         {
             var game = await this.recruitmentRepository.GetByIdAsync(id);
+
+            if (game == null)
+            {
+                throw new ServiceException(ServiceException.InvalidRequestParameters);
+            }
+
             this.recruitmentRepository.Delete(game);
             await this.recruitmentRepository.SaveChangesAsync();
+        }
+
+        public async Task<Recruitment> GetRecruitmentAsync(string id)
+        {
+            return await this.recruitmentRepository.GetByIdAsync(id);
         }
     }
 }
